refactor: Improve client deletion API (#4261)

+++ core/src/main/client/ClientService.ts
@@ -47,27 +47,35 @@ export class ClientService {
     this.backend = new ClientBackendRepository(this.apiClient);
   }
 
-  public deleteLocalClient(): Promise<string> {
-    return this.database.deleteLocalClient();
-  }
-
   public getClients(): Promise<RegisteredClient[]> {
     return this.backend.getClients();
   }
 
   /**
    * Will delete the given client from backend and will also delete it from the local database
+   *
+   * note: use deleteLocalClient if you wish to delete the client currently used by the user
+   *
    * @param clientId The id of the client to delete
-   * @param password Password of the owning user
-   * @param isLocalClient Is the client also the client currently used by the app (as opposed to a client the user owns but not currenly in use)
-   * @returns Promise
+   * @param password? Password of the owning user. Can be omitted for temporary devices
    */
-  public async deleteClient(clientId: string, password: string, isLocalClient: boolean = false): Promise<unknown> {
+  public async deleteClient(clientId: string, password?: string): Promise<unknown> {
     const userId: QualifiedId = {id: this.apiClient.userId as string, domain: this.apiClient.domain || ''};
     await this.backend.deleteClient(clientId, password);
-    return isLocalClient
-      ? this.database.deleteLocalClient()
-      : this.database.deleteClient(this.cryptographyService.constructSessionId(userId, clientId));
+    return this.database.deleteClient(this.cryptographyService.constructSessionId(userId, clientId));
+  }
+
+  /**
+   * Will delete the local client (client currently in use by the user) from backend and will also delete it from the local database
+   * @param password? Password of the owning user. Can be omitted for temporary devices
+   */
+  public async deleteLocalClient(password?: string): Promise<string> {
+    const localClientId = this.apiClient.context?.clientId;
+    if (!localClientId) {
+      throw new Error('Trying to delete local client, but local client has not been set');
+    }
+    await this.backend.deleteClient(localClientId, password);
+    return this.database.deleteLocalClient();
   }
 
   public getLocalClient(): Promise<MetaClient> {
